Django-signals_orm-0x04/requirements.txt
django
djangorestframework
pillow
django-environ
djangorestframework-simplejwt
django-filter
isort
